+++ src/timer.cc
@@ -2,7 +2,7 @@
 
 Timer::Timer()
 {
-    duration_ = 0;
+    duration_ = 0.0;
     then_ = 0;
     now_ = 0;
 }
@@ -12,7 +12,7 @@ Timer::~Timer()
 
 }
 
-void Timer::SetInterval(uint32_t ms)
+void Timer::SetInterval(double ms)
 {
     duration_ = ms;
     then_ = SDL_GetTicks();
@@ -21,7 +21,7 @@ void Timer::SetInterval(uint32_t ms)
 bool Timer::Tick()
 {
     now_ = SDL_GetTicks();
-    uint32_t delta = now_ - then_;
+    double delta = now_ - then_;
     if (delta >= duration_)
     {
         then_ = now_;
